refactor(dev): refactor types locations
